requirements.txt
Flask==2.3.2
rawpy==0.25.1
numpy==1.26.4
opencv-python==4.8.0.76
Pillow==10.0.0
psutil==5.9.5
